rename "adapter" to "fmt"
